requirements.txt
aiogram
aiosqlite
apscheduler
python-dotenv
holidays